drivers: media/camera: Overwrite the camera_v2 by the stock kernel
For now, I have to overwrite all the camera_v2 files or I cannot revive
the camera. Sad that the new patches are not merged on the camera
driver.

I tried to use the latest camera_v2 (excepts ./sensor directory) files with only
necessary Samsung codes from the stock firmware but failed.

This is a workaround, but seems working well with the CAF kernel base.

Signed-off-by: Deokgyu Yang <secugyu@gmail.com>